feat(chat):导入了相关的包，添加了雪花算法、消息转换器等工具类，添加了几个较为简单crud的api

+++ src/main/java/com/removel/accp/model/entity/ChatMessage.java
@@ -20,7 +20,7 @@ public class ChatMessage {
 
     // 标记这个信息归属的对话id
     @TableField("session_id")
-    private Integer sessionId;
+    private Long sessionId;
 
     // 这个信息的内容
     @TableField("content")
